check out /newmenu.html
Experimenting with a using a table to hold the menu. Easier for line
items and fits nicely into columns. Its a pain manually typing all of
his menu into this but it should look a lot nicer when finished. Using
a pic of the current menu off of yelp for reference, since menu on
website is atrocious